Separate Game and GameLauncher classes
