+++ lutris/gui/views/base.py
@@ -7,6 +7,7 @@ from lutris.database.games import get_game_for_service
 from lutris.database.services import ServiceGameCollection
 from lutris.game import Game
 from lutris.game_actions import GameActions, get_game_actions
+from lutris.game_launcher import GameLauncher
 from lutris.gui.widgets.contextual_menu import ContextualMenu
 from lutris.gui.widgets.utils import MEDIA_CACHE_INVALIDATED
 from lutris.util.log import logger
@@ -38,7 +39,7 @@ class GameView:
         self.connect("button-press-event", self.popup_contextual_menu)
         self.connect("key-press-event", self.handle_key_press)
 
-        self.game_start_hook_id = GObject.add_emission_hook(Game, "game-start", self.on_game_start)
+        self.game_start_hook_id = GObject.add_emission_hook(GameLauncher, "game-start", self.on_game_start)
 
     def set_game_store(self, game_store):
         self.game_store = game_store
@@ -150,7 +151,7 @@ class GameView:
     def get_game_id_for_path(self, path):
         raise NotImplementedError()
 
-    def on_game_start(self, game):
+    def on_game_start(self, game_launcher: GameLauncher):
         """On game start, we trigger an animation to show the game is starting; it runs at least
         one cycle, but continues until the game exits the STATE_LAUNCHING state."""
 
@@ -181,8 +182,8 @@ class GameView:
             if elapsed > cycle_time:
                 # Check for stopping and pausing only at cycle end, so we don't do it too often,
                 # and to avoid a janky looking visible snap-back to full size.
-                if game.state != game.STATE_LAUNCHING:
-                    if self.image_renderer.inset_game(game.id, 0.0):
+                if game_launcher.state != game_launcher.STATE_LAUNCHING:
+                    if self.image_renderer.inset_game(game_launcher.game.id, 0.0):
                         self.queue_draw()
                     return False
 
@@ -201,7 +202,7 @@ class GameView:
             else:
                 fraction = max_indent * (cycle * 2 / cycle_time)
 
-            if self.image_renderer.inset_game(game.id, fraction):
+            if self.image_renderer.inset_game(game_launcher.game.id, fraction):
                 self.queue_draw()
 
             return True  # Return True to call again after another timeout
